refactor(cli): rename installer to kubesoloctl management CLI
Restructure the Go installer into a multi-command CLI tool following
the Cilium CLI pattern. Thin cmd/kubesoloctl/main.go entry point
delegates to internal/cli package which owns the full cobra command
tree.

Commands: install, uninstall, upgrade, download, check, kubeconfig,
reset, completion, version. Running without a subcommand defaults to
install for backward compatibility.

- Rename cmd/installer → cmd/kubesoloctl
- Rename internal/installer → internal/cli
- Split monolithic main.go into per-command files
- Add Makefile targets: build-kubesoloctl, test-kubesoloctl
- Update scripts/get.sh and CI workflow for new binary name
- Bump default version to v1.1.5
