+++ Makefile
@@ -215,47 +215,45 @@ archive:
 archive-musl:
 	tar -czf dist/kubesolo-musl.tar.gz dist/kubesolo install.sh
 
-# ── Installer binary ──────────────────────────────────────────────────────────
+# ── kubesoloctl binary ──────────────────────────────────────────────────────────
 #
-# The installer is built with CGO_ENABLED=0 (pure Go). A single binary per
+# kubesoloctl is built with CGO_ENABLED=0 (pure Go). A single binary per
 # architecture runs on both glibc and musl systems, so there is no libc split.
 #
 # Supported targets: linux/amd64, linux/arm64, linux/arm (armhf), linux/riscv64
 
-INSTALLER_LDFLAGS = -s -w \
+KUBESOLOCTL_LDFLAGS = -s -w \
 	-X main.Version=$(VERSION) \
 	-X main.Commit=$(COMMIT) \
 	-X main.BuildDate=$(BUILD_DATE)
 
-INSTALLER_OUTPUT ?= ./dist/installer-$(GOOS)-$(GOARCH)
+KUBESOLOCTL_OUTPUT ?= ./dist/kubesoloctl-$(GOOS)-$(GOARCH)
 
-# Build the installer for the current GOOS/GOARCH.
-# Pass GOARM=7 when targeting arm (armhf/ARMv7), e.g.: make build-installer GOARCH=arm GOARM=7
-.PHONY: build-installer
-build-installer:
-	@mkdir -p $(dir $(INSTALLER_OUTPUT))
+# Build kubesoloctl for the current GOOS/GOARCH.
+# Pass GOARM=7 when targeting arm (armhf/ARMv7), e.g.: make build-kubesoloctl GOARCH=arm GOARM=7
+.PHONY: build-kubesoloctl
+build-kubesoloctl:
+	@mkdir -p $(dir $(KUBESOLOCTL_OUTPUT))
 	CGO_ENABLED=0 GOOS=$(GOOS) GOARCH=$(GOARCH) GOARM=$(GOARM) go build \
-		-ldflags="$(INSTALLER_LDFLAGS)" \
-		-o $(INSTALLER_OUTPUT) \
-		./cmd/installer
+		-ldflags="$(KUBESOLOCTL_LDFLAGS)" \
+		-o $(KUBESOLOCTL_OUTPUT) \
+		./cmd/kubesoloctl
 
-# Build the installer for all supported architectures
-.PHONY: build-installer-all
-build-installer-all:
-	GOARCH=amd64   INSTALLER_OUTPUT=./dist/installer-linux-amd64   make build-installer
-	GOARCH=arm64   INSTALLER_OUTPUT=./dist/installer-linux-arm64   make build-installer
-	GOARCH=arm     INSTALLER_OUTPUT=./dist/installer-linux-arm     make build-installer GOARM=7
-	GOARCH=riscv64 INSTALLER_OUTPUT=./dist/installer-linux-riscv64 make build-installer
+# Build kubesoloctl for all supported architectures
+.PHONY: build-kubesoloctl-all
+build-kubesoloctl-all:
+	GOARCH=amd64   KUBESOLOCTL_OUTPUT=./dist/kubesoloctl-linux-amd64   make build-kubesoloctl
+	GOARCH=arm64   KUBESOLOCTL_OUTPUT=./dist/kubesoloctl-linux-arm64   make build-kubesoloctl
 
-# Run installer tests (no CGO required, no cross-compiler needed)
-.PHONY: test-installer
-test-installer:
-	CGO_ENABLED=0 go test ./internal/installer/... -v -count=1
+# Run kubesoloctl tests (no CGO required, no cross-compiler needed)
+.PHONY: test-kubesoloctl
+test-kubesoloctl:
+	CGO_ENABLED=0 go test ./internal/cli/... -v -count=1
 
-# Clean installer build artefacts
-.PHONY: clean-installer
-clean-installer:
-	rm -f ./dist/installer-*
+# Clean kubesoloctl build artefacts
+.PHONY: clean-kubesoloctl
+clean-kubesoloctl:
+	rm -f ./dist/kubesoloctl-*
 
 # Include custom make targets
 -include $(wildcard .dev/*.make)
