+++ cmd/kubesoloctl/main.go
@@ -1,0 +1,25 @@
+// cmd/kubesoloctl is the management CLI for KubeSolo — a single-node
+// Kubernetes distribution for constrained edge environments.
+//
+// It is built with CGO_ENABLED=0 so a single binary per architecture covers
+// both glibc and musl systems.
+package main
+
+import (
+	"os"
+
+	"github.com/portainer/kubesolo/internal/cli"
+)
+
+var (
+	Version   = "dev"
+	BuildDate = "unknown"
+	Commit    = "unknown"
+)
+
+func main() {
+	cli.SetVersionInfo(Version, Commit, BuildDate)
+	if err := cli.Execute(); err != nil {
+		os.Exit(1)
+	}
+}
